HBASE-7808 Refactor Store to use HRegionFileSystem
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1452936 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/Store.java
@@ -145,7 +145,7 @@ public interface Store extends HeapSize, StoreConfigInformation {
    * @param isCompaction whether we are creating a new file in a compaction
    * @return Writer for a new StoreFile in the tmp dir.
    */
-  public StoreFile.Writer createWriterInTmp(int maxKeyCount,
+  public StoreFile.Writer createWriterInTmp(long maxKeyCount,
     Compression.Algorithm compression, boolean isCompaction) throws IOException;
 
   // Compaction oriented methods
